python_service/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
pydantic-settings
fastapi-mail
python-dotenv
psycopg2-binary
requests
